refactor: rename launch-local to launch-platform
