+++ lib/cmds/run.js
@@ -20,7 +20,7 @@ const fs = require('fs');
 const path = require('path');
 const crypto = require('crypto');
 const os = require('os');
-const createLauncher = require('../launch-local');
+const createLauncher = require('../launch-platform');
 const createConsoleReporter = require('../../console-reporter');
 
 const moment = require('moment');
@@ -40,6 +40,7 @@ class RunCommand extends Command {
     // Collect all input files for reading/parsing - via args, --config, or -i
     const inputFiles = argv.concat(flags.input || [], flags.config || []);
 
+    // TODO: Move into PlatformLocal
     if (flags.dotenv) {
       const dotEnvPath = path.resolve(process.cwd(), flags.dotenv);
       try {
@@ -69,10 +70,21 @@ class RunCommand extends Command {
         runnerOpts.count = 1;
       }
 
+      let platformConfig = {};
+      if (flags['platform-opt']) {
+        for(const opt of flags['platform-opt']) {
+          const [k, v] = opt.split('=');
+          platformConfig[k] = v;
+        }
+      }
+
       const launcherOpts = {
         platform: flags.platform,
+        platformConfig,
         mode: flags.platform === 'local' ? 'distribute' : 'multiply',
         region: 'eu-west-1', // TODO: Read from --platform-option region=eu-west-1
+        count: parseInt(flags.count || 1, 10),
+        cliArgs: flags,
       };
 
       let launcher = await createLauncher(
@@ -265,7 +277,7 @@ RunCommand.flags = {
     char: 'i',
     description: 'Input script file',
     multiple: true,
-    hidden: true
+    hidden: true,
   }),
   solo: flags.boolean({
     char: 's',
@@ -277,7 +289,14 @@ RunCommand.flags = {
   platform: flags.string({
     description: 'Runtime platform',
     default: 'local',
-    hidden: true,
+  }),
+  'platform-opt': flags.string({
+    description: 'Set a platform-specific option, e.g. --platform region=eu-west-1 for AWS Lambda',
+    multiple: true,
+  }),
+  count: flags.string({
+    // locally defaults to number of CPUs with mode = distribute
+    default: '1',
   })
 };
 
@@ -377,6 +396,9 @@ async function sendTelemetry(script, flags) {
       properties.targetHash = targetHash;
     }
 
+    properties.platform = flags.platform;
+    properties.count = flags.count;
+
     if (properties.targetHash) {
       properties.distinctId = properties.targetHash;
     }
@@ -424,7 +446,9 @@ async function sendTelemetry(script, flags) {
 
     for (const scenario of script.scenarios || []) {
       if (OSS_ENGINES.indexOf(scenario.engine || 'http') > -1) {
-        properties.enginesUsed.push(scenario.engine || 'http');
+        if (properties.enginesUsed.indexOf(scenario.engine || 'http') === -1) {
+          properties.enginesUsed.push(scenario.engine || 'http');
+        }
       }
     }
