+++ lib/launch-platform.js
