+++ lib/platform/local/worker.js
@@ -110,7 +110,7 @@ async function prepare(opts) {
 
   // NOTE: We don't subscribe plugins to stats/done events from
   // individual runner instances here - those are handled in
-  // launch-local instead. (If we subscribe plugins to events here,
+  // launch-platform instead. (If we subscribe plugins to events here,
   // they will receive individual stats/done events from workers,
   // instead of objects that have been properly aggregated.)
   const stubEE = new EventEmitter();
