+++ test/core/http/test_function.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { test } = require('tap');
-const createRunner = require('../../../lib/launch-local');
+const createRunner = require('../../../lib/launch-platform');
 const sleep = require('../../../lib/util/sleep');
 const path = require('path');
